Initial commit
This is intended to be a test application. Different branches will
be instrumented using:

- Honeycomb
- Datadog
- Circonus

and so on. The purpose of this is to get a feel for how an application
using those services will look, and get some silly data into those
services so that we can try to understand how we might use them.